+++ nucypher-core-python/nucypher_core/umbral.py
@@ -1,4 +1,4 @@
-from .._nucypher_core import umbral as _umbral
+from ._nucypher_core import umbral as _umbral
 
 SecretKey = _umbral.SecretKey
 PublicKey = _umbral.PublicKey
@@ -6,6 +6,7 @@ SecretKeyFactory =_umbral.SecretKeyFactory
 Signature = _umbral.Signature
 Signer = _umbral.Signer
 Capsule = _umbral.Capsule
+KeyFrag = _umbral.KeyFrag
 VerifiedKeyFrag = _umbral.VerifiedKeyFrag
 CapsuleFrag = _umbral.CapsuleFrag
 VerifiedCapsuleFrag = _umbral.VerifiedCapsuleFrag